+++ bin/kube-include.sh
@@ -24,17 +24,17 @@ else
 fi
 
 # Supported versions of SAS Viya 4
-# Updated: 23MAR23
-# 2022.1  LTS 1.20 1.22
+# Updated: 18MAY23
 # 2022.09 LTS 1.21 1.24
-# 2022.12     1.22 1.24
-# 2023.01     1.22 1.24
+# 2023.03 LTS 1.23 1.25
 # 2023.02     1.22 1.24
 # 2023.03     1.23 1.25
-if [[ $KUBE_SERVER_VER =~ v1.2[0-9] ]]; then
+# 2023.04     1.23 1.25
+# 2023.05     1.24 1.26
+if [[ $KUBE_SERVER_VER =~ v1.2[1-9] ]]; then
   :
 else 
-  log_warn "The detected version of Kubernetes [$KUBE_SERVER_VER] is not supported by any of the currently supported releases of SAS Viya 4."
+  log_warn "The detected version of Kubernetes [$KUBE_SERVER_VER] is not supported by any of the currently supported releases of SAS Viya."
   log_warn "This script might not work as expected. Support might not be available until Kubernetes is upgraded to a supported version."
 fi