+++ logging/bin/getlogs.py
@@ -10,11 +10,12 @@ import os, sys
 import json, csv
 
 ## Cleanup temp file if not previously deleted, Create target directory
-if not os.path.isdir("GetLogs_Files"):
-    os.mkdir("GetLogs_Files")
+DEFAULT_DIR = "GetLogsFiles"
+if not os.path.isdir(DEFAULT_DIR):
+    os.mkdir(DEFAULT_DIR)
 
-if os.path.isfile("GetLogs_Files/temp_SAS_OSQ.TXT"):
-    os.remove("GetLogs_Files/temp_SAS_OSQ.TXT")
+if os.path.isfile(DEFAULT_DIR + "/temp_SAS_OSQ.TXT"):
+    os.remove(DEFAULT_DIR + "/temp_SAS_OSQ.TXT")
 
 def validate_input(dict):
     """Validate the arguments passed by the user to ensure script will function properly"""
@@ -26,8 +27,8 @@ def validate_input(dict):
     if dict['create-auth']:
         if (type(dict['create-auth'] == list)):
             dict['create-auth'] = " ".join(dict['create-auth'])
-        if (not "GetLogs_Files/" in  dict['create-auth']):
-            dict['create-auth'] = os.path.join("GetLogs_Files", dict['create-auth'])
+        if (not (DEFAULT_DIR + "/") in  dict['create-auth']):
+            dict['create-auth'] = os.path.join(DEFAULT_DIR, dict['create-auth'])
         x = open(dict['create-auth'], "w")
         x.write('{"username": " " , "password": " " , "host": " ", "port": " "}')
         print("\nAuth File created.\n")
@@ -35,8 +36,8 @@ def validate_input(dict):
     if dict['auth-file']:
         if (type(dict['auth-file'] == list)):
             dict['auth-file'] = " ".join(dict['auth-file'])
-        if (not "GetLogs_Files/" in  dict['auth-file']):
-            dict['auth-file'] = os.path.join("GetLogs_Files",  dict['auth-file'])
+        if (not (DEFAULT_DIR + "/") in  dict['auth-file']):
+            dict['auth-file'] = os.path.join(DEFAULT_DIR,  dict['auth-file'])
         try: 
             a = open(dict['auth-file'], 'r')
             authDict = json.loads(a.read())
@@ -91,8 +92,8 @@ def validate_input(dict):
                     print('Setting default filetype: CSV')
                     dict['out-filename'] = dict['out-filename'] + ".csv"
 
-        if (not "GetLogs_Files/" in dict['out-filename']):
-            dict['out-filename'] = os.path.join("GetLogs_Files", dict['out-filename'])
+        if (not (DEFAULT_DIR + "/") in dict['out-filename']):
+            dict['out-filename'] = os.path.join(DEFAULT_DIR, dict['out-filename'])
         
         if os.path.isfile(dict['out-filename']):
             if (dict['force'] == False):
@@ -107,12 +108,12 @@ def validate_input(dict):
         if ((not ".json" in dict['savequery'])):
             print('Error: Not a supported filetype for the saved query file.')
             exit()
-        if (not "GetLogs_Files/" in dict['savequery']):
-            dict['savequery']= os.path.join("GetLogs_Files", dict['savequery'])
+        if (not (DEFAULT_DIR + "/") in dict['savequery']):
+            dict['savequery']= os.path.join(DEFAULT_DIR, dict['savequery'])
 
     if dict['query-filename']:
-        if (not "GetLogs_Files/" in dict['query-filename']):
-            dict['query-filename'] = os.path.join("GetLogs_Files", dict['query-filename'])
+        if (not (DEFAULT_DIR + "/") in dict['query-filename']):
+            dict['query-filename'] = os.path.join(DEFAULT_DIR, dict['query-filename'])
         
         if (not os.path.isfile(dict['query-filename'])):
             print("Error: Invalid query file path.")
@@ -145,8 +146,8 @@ def build_query(dict): ##Generates Query using Opensearch DSL
     first = True 
     argcounter=0    ##Counts unique options entered by user, sets min_match to this number
     if (not dict['query-filename']):     ##If User has not specified query file, create temp file for one.                                                                                                                                                                                                                                                                                                                                                                                   first = True
-        x = open("GetLogs_Files/temp_SAS_OSQ.TXT", 'x')
-        x = open("GetLogs_Files/temp_SAS_OSQ.TXT", 'wt')  ## Create txt file to write query, make it writable
+        x = open(DEFAULT_DIR + "/temp_SAS_OSQ.TXT", 'x')
+        x = open(DEFAULT_DIR + "/temp_SAS_OSQ.TXT", 'wt')  ## Create txt file to write query, make it writable
         x.write('{"size": ' + str(dict['maxInt']) + ',"sort": [{"@timestamp": {"order": "desc","unmapped_type": "boolean"} }]')     ## Establish size of query, remove scoring
         x.write(', "query": {"bool": {"must":[ {"range": {"@timestamp": {"gte": "' + dict['dateTimeStart'] + '","lt": "' + dict['dateTimeEnd'] + '"} } }], ')   ##Establish Query with Range Requirements, 3 open brackets
         x.write('"should": [ ')     ## Should Clause, search results must inlcude at least one of each specified option
@@ -181,10 +182,10 @@ def build_query(dict): ##Generates Query using Opensearch DSL
         if (first==False):
             x.write(']')
         x.write('} } }') ##Close 3 open brackets, end query
-        x = open("GetLogs_Files/temp_SAS_OSQ.TXT", 'rt')
+        x = open(DEFAULT_DIR + "/temp_SAS_OSQ.TXT", 'rt')
         query =  " ".join([line.strip() for line in x])
         x.close()
-        os.remove("GetLogs_Files/temp_SAS_OSQ.TXT") ##Remove Temp File Created for Query Generation
+        os.remove(DEFAULT_DIR + "/temp_SAS_OSQ.TXT") ##Remove Temp File Created for Query Generation
     else:
         x = open(dict['query-filename'], 'rt')
         query =  " ".join([line.strip() for line in x]) ## Turn file into string, return.
@@ -195,7 +196,9 @@ def build_query(dict): ##Generates Query using Opensearch DSL
 ##List of VALID arguments that are read from user as soon as program is run, nargs=+ indicates that argument takes multiple whitespace separated values. 
 def get_arguments():
     """Defines the arguments a user can pass and parses them"""
-    parser = argparse.ArgumentParser(prog='getLogs.py', usage='\n%(prog)s [options]', description="This program generates OpenSearch DSL Queries from user specified parameters, and submits them to a database to retrieve logs. The flags below provide specifications for your Query, and can be placed in any order. \n     IMPORTANT NOTES: \n -Connection settings are required in order to run the program. You can create config files to autofill this using -cf, and call them using -af\n    -The NAMESPACE*, POD*, CONTAINER*, LOGSOURCE* and LEVEL* options accept multiple, space-separated, values (e.g. --level INFO NONE) Please refrain from passing single quotes ('') into arguments. \n    -All Generated Program files are placed in GetLogs_Files where the program is run. It is not required to include this directory when passing created files in as arguments\n    -Correct time format is Y-M-D H:M:S. Ex: 1999-02-07 10:00:00 \n     -All default values for username, password, host, and port are derived from the ENV variables ESUSER, USPASSWD, ESHOST, ESPORT", formatter_class=argparse.RawTextHelpFormatter)
+    parser = argparse.ArgumentParser(prog='getLogs.py', usage='\n%(prog)s [options]', description="This program generates OpenSearch DSL Queries from user specified parameters, and submits them to a database to retrieve logs. The flags below provide specifications for your Query, and can be placed in any order. \n   IMPORTANT NOTES: \n    *Connection settings are required in order to run the program. You can create config files to autofill this using --create-auth, and call them using --auth-file \n    *The NAMESPACE*, POD*, CONTAINER*, LOGSOURCE* and LEVEL* options accept multiple, space-separated, values (e.g. --level INFO NONE) Please refrain from passing single quotes ('') into arguments. \n    *All Generated Program files are placed in a folder titled GetLogssFiles in the directory where the program is run. It is not required to include this directory when passing created files in as arguments.\n    *Correct time format is Y-M-D H:M:S. Ex: 1999-02-07 10:00:00 \n    *All default values for username, password, host, and port are derived from the ENV variables ESUSER, USPASSWD, ESHOST, ESPORT\n\n\n \t\t\t\t QUERY SEARCH PARAMETERS: \n\n", formatter_class=argparse.RawTextHelpFormatter)
+    
+    ##Search Params
     parser.add_argument('-n', '--namespace', required=False, dest="kube.namespace", nargs='*', metavar="NAMESPACE", help="\nOne or more Viya deployments/Kubernetes Namespace for which logs are sought\n\n")
     parser.add_argument('-nx', '--namespace-exclude', required=False, dest="kube.namespace-ex", nargs='*', metavar="NAMESPACE", help='\nOne or more namespaces for which logs should be excluded from the output\n\n')
     parser.add_argument('-p', '--pod', required=False, dest="kube.pod", nargs='*', metavar="POD", help='\nOne or more pods for which logs are sought\n\n')
@@ -206,7 +209,9 @@ def get_arguments():
     parser.add_argument('-sx', '--logsource-exclude', required=False, dest="logsource-ex",nargs='*', metavar="LOGSOURCE",  help = "\nOne or more logsource for which logs should be excluded from the output\n\n")
     parser.add_argument('-l', '--level', required=False, dest='level', nargs='*', metavar="LEVEL",  help = "\nOne or more message levels for which logs are sought\n\n")
     parser.add_argument('-lx', '--level-exclude', required=False, dest = 'level-exclude', nargs='*', metavar="LEVEL", help = "\nOne or more message levels for which logs should be excluded from the output.\n\n")
-    parser.add_argument('-se', '--search', required=False, dest= "message", nargs='*', metavar="MESSAGE",  help = "\nWord or phrase contained in log message. Do not include single quotes ('')\n\n")
+    parser.add_argument('-se', '--search', required=False, dest= "message", nargs='*', metavar="MESSAGE",  help = "\nWord or phrase contained in log message. Do not include single quotes ('')\n\n\n \t\t\t QUERY OUTPUT SETTINGS: \n\n")
+
+    ##Query and Output Params
     parser.add_argument('-m', '--maxrows', required=False, dest ="maxInt", type=int, metavar="INTEGER", default=0,  help = "\nThe maximum number of log messsages to return. Max possible rows is 10000\n\n")
     parser.add_argument('-q', '--query-file ', required=False, dest="query-filename", metavar="FILENAME.*", help = "\nName of file containing search query (Including filetype) at end. Program will submit query from file, ALL other query parmeters ignored. Supported filetypes: .txt, .json\n\n")
     parser.add_argument('-sh', '--show-query', required=False, dest="showquery", action= "store_true", help = "\n Display example of actual query that will be submitted during execution.\n\n")
@@ -216,13 +221,15 @@ def get_arguments():
     parser.add_argument('-f','--force',  required=False, dest="force", action= "store_true", help = "\n If this option is provided, the output results file will be overwritten if it already exists.\n\n")
     parser.add_argument('-fi','--fields',  required=False, dest="fields", nargs="*", metavar= "FIELDS", default=['@timestamp', 'level', 'logsource', 'namespace','pod', 'container', 'message'], help = "\n Specify output columns (CSV file only) Please provide a space separated list of fields. \n Default fields: @timestamp level logsource namespace pod container message \n Additional arguments: host index properties debug \n\n")
     parser.add_argument('-st', '--start', required=False, dest="dateTimeStart", nargs='*', metavar="DATETIME",  default = time.strftime("%Y-%m-%d %H:%M:%S", time.localtime(time.mktime(time.localtime()) - 3600)), help = "\nDatetime for start of period for which logs are sought (default: 1 hour ago).\n\n")
-    parser.add_argument('-en', '--end', required=False, dest="dateTimeEnd",nargs='*', metavar="DATETIME",  default = time.strftime("%Y-%m-%d %H:%M:%S", time.localtime()), help = "\nDatetime for end of period for which logs are sought (default: now). \n\n")
-    parser.add_argument('-i', '--index', required=False, dest="index", metavar="INDEX", default="viya_logs-*", help = "\nDetermine which index to perform the search in. Default: viya-logs-*\n\n")
+    parser.add_argument('-en', '--end', required=False, dest="dateTimeEnd",nargs='*', metavar="DATETIME",  default = time.strftime("%Y-%m-%d %H:%M:%S", time.localtime()), help = "\nDatetime for end of period for which logs are sought (default: now). \n\n\n \t\t\t CONNECTION SETTINGS: \n\n")
+    
+    parser.add_argument('-i', '--index', required=False, dest="index", metavar="INDEX", default="viya_logs-*") ## help = "\nDetermine which index to perform the search in. Default: viya-logs-*\n\n
+    ##Connection settings
     parser.add_argument('-us','--user',  required=False, dest="userName", default=os.environ.get("ESUSER"), help = "\nUsername for connecting to OpenSearch/Kibana (default: $ESUSER)\n\n")
     parser.add_argument('-pw', '--password', required=False,  dest="password", default=os.environ.get("ESPASSWD"), help = "\nPassword for connecting to OpenSearch/Kibana  (default: $ESPASSWD)\n\n")
     parser.add_argument('-ho', '--host', required=False,  dest="host", default=os.environ.get("ESHOST"), help = "\nHostname for connection to OpenSearch/Kibana (default: $ESHOST)\n\n")
     parser.add_argument('-po', '--port', required=False,  dest="port", default=os.environ.get("ESPORT"), help = "\nPort number for connection to OpenSearch/Kibana (default: $ESPORT)\n\n")
-    parser.add_argument('-ss', '--disable-ssl', required=False, dest = "ssl", action= "store_false", help = "\n If this option is provided, SSL will not be used to connect to the database.\n\n")
+    parser.add_argument('-nossl', '--disable-ssl', required=False, dest = "ssl", action= "store_false", help = "\n If this option is provided, SSL will not be used to connect to the database.\n\n")
     parser.add_argument('-af', '--auth-file', required=False, dest="auth-file", metavar="FILENAME.*", nargs='*', help = "\n The program will read the connection information from this file (include filetype).\n\n")
     parser.add_argument('-cf', '--create-auth', required=False, dest="create-auth", metavar="FILENAME.*", nargs='*', help = "\n Creates a template authentication file (include filetype) under the specified file name. Connection information can be entered here.")
     return parser.parse_args().__dict__