+++ logging/bin/onboard.sh
@@ -16,12 +16,12 @@ this_script=`basename "$0"`
 function show_usage {
   log_message  "Usage: $this_script --namespace NAMESPACE [--tenant TENANT] [OPTIONS]"
   log_message  ""
-  log_message  "'Onboards' a Viya deployment (namespace) or a specific tenant within that deployment.  This process allows admins responsible for a Viya deployment (or a single tenant within a given deployment) to work with log messages collected from the deployment (or tenant within the deployment)."
+  log_message  "'Onboards' a SAS Viya deployment (namespace) or a specific tenant within that deployment.  This process allows admins responsible for a SAS Viya deployment (or a single tenant within a given deployment) to work with log messages collected from the deployment (or tenant within the deployment)."
   log_message  "The onboarding process creates the security access controls and, optionally, an initial user granted access.  In addition, an OpenSearch Dashboards tenant space is created and populated with an initial set of OpenSearch Dashboards content (e.g. visualizations, dashboards, etc.)."
   log_message  ""
   log_message  "    Arguments:"
-  log_message  "     -ns, --namespace   NAMESPACE - (Required) The Viya deployment/Kubernetes Namespace to 'on-board' or the namespace in which the tenant to 'on-board' resides."
-  log_message  "     -t,  --tenant      TENANT    - (Optional) The tenant within the specified Viya deployment/Kubernetes Namespace to 'on-board'."
+  log_message  "     -ns, --namespace   NAMESPACE - (Required) The SAS Viya deployment/Kubernetes Namespace to 'on-board' or the namespace in which the tenant to 'on-board' resides."
+  log_message  "     -t,  --tenant      TENANT    - (Optional) The tenant within the specified SAS Viya deployment/Kubernetes Namespace to 'on-board'."
   log_message  ""
   log_message  "    Options:"
   log_message  "        -u, --user [USER]           - Create an initial user with access to this OpenSearch Dashboards tenant space. User name is optional, by default its name will combine the OpenSearch Dashboards tenant space name with '_admin'."
@@ -137,10 +137,10 @@ fi
 ktenant=$nst
 
 if [ -n "$tenant" ]; then
-   tenant_description="An OpenSearch Dashboards tenant space for tenant [$tenant] within Viya deployment (namespace) [$namespace]."
+   tenant_description="An OpenSearch Dashboards tenant space for tenant [$tenant] within SAS Viya deployment (namespace) [$namespace]."
    log_notice "Onboarding tenant [$tenant] within namespace [$namespace] [$(date)]"
 else
-   tenant_description="An OpenSearch Dashboards tenant space for Viya deployment (namespace) [$namespace]."
+   tenant_description="An OpenSearch Dashboards tenant space for SAS Viya deployment (namespace) [$namespace]."
    log_notice "Onboarding namespace [$namespace] [$(date)]"
 fi